+++ packages/components/src/components/HandleMessage.js
@@ -11,8 +11,8 @@ import { MethodGenerator } from './MethodGenerator';
  */
 const websocketHandleMessageConfig = {
   python: {
-    methodLogic: `if len(self.message_handlers) == 0:
-  print("\\033[94mReceived raw message:\\033[0m", message)
+    methodLogic: String.raw`if len(self.message_handlers) == 0:
+  print("\033[94mReceived raw message:\033[0m", message)
 else:
     for handler in self.message_handlers:
       handler(message)`